refactor: turn event page to be a server-side component
